Fixes to warnings
Removed everything in sparse_vector_float and sparse_vector_float_serial, and t.cpp that caused a warning, fixed an error in a test, and removed the tempTests file
