--- tests/stress/sparse_vec_float_tempTests.cpp
@@ -1,1009 +1,0 @@
-#include <iostream>
-#include <fstream>
-#include <cassert>
-
-#include <bm.h>
-#include <bmsparsevec_float.h>
-#include <bmsparsevec_float_serial.h>
-#include <bmsparsevec_algo.h>
-#include <random>
-
-typedef bm::sparse_vector_float<bm::sparse_vector<unsigned int, bm::bvector<>>> sparseVecFloat;
-
-//----------------------------------------------------------------
-
-void SparseVecFloatConstIteratorTests(){
-    
-    float toAdd[] = {1.0123, 2.468, 340000.56};
-
-    sparseVecFloat testSVF;
-    testSVF.import(toAdd, 3);
-
-    auto floatEq = [](float a, float b) {
-        return std::fabs(a - b) < 0.001f;
-    };
-    
-    // --- construction ---
-    sparseVecFloat::const_iterator defaultit;
-    sparseVecFloat::const_iterator itFromSV(&testSVF);
-    sparseVecFloat::const_iterator itFromSVPos(&testSVF, 1);
-    sparseVecFloat::const_iterator itBegin = testSVF.begin();
-    sparseVecFloat::const_iterator itEnd   = testSVF.end();
-    sparseVecFloat::const_iterator itCopy(itBegin);
-    
-    // --- operator* and value() ---
-    assert(floatEq(*itBegin, toAdd[0]));
-    assert(floatEq(itBegin.value(), toAdd[0]));
-    assert(floatEq(*itFromSVPos, toAdd[1]));
-
-    // --- valid() ---
-    assert(itBegin.valid());
-    assert(!itEnd.valid());
-    
-    // --- invalidate() ---
-    sparseVecFloat::const_iterator itInvalid = testSVF.begin();
-    itInvalid.invalidate();
-    assert(!itInvalid.valid());
-
-    // --- pos() ---
-    assert(itBegin.pos() == 0);
-    assert(itFromSVPos.pos() == 1);
-
-    // --- operator== and operator!= ---
-    sparseVecFloat::const_iterator itA = testSVF.begin();
-    sparseVecFloat::const_iterator itB = testSVF.begin();
-    assert(itA == itB);
-    assert(!(itA != itB));
-    assert(itA != itEnd);
-    assert(!(itA == itEnd));
-
-    // --- operator< <= > >= ---
-    sparseVecFloat::const_iterator itFirst  = testSVF.begin();
-    sparseVecFloat::const_iterator itSecond(&testSVF, 1);
-    assert(itFirst  <  itSecond);
-    assert(itFirst  <= itSecond);
-    assert(itFirst  <= itFirst);
-    assert(itSecond >  itFirst);
-    assert(itSecond >= itFirst);
-    assert(itFirst  >= itFirst);
-
-    // --- prefix operator++ ---
-    sparseVecFloat::const_iterator itPre = testSVF.begin();
-    ++itPre;
-    assert(itPre.pos() == 1);
-    assert(floatEq(*itPre, toAdd[1]));
-    ++itPre;
-    assert(itPre.pos() == 2);
-    assert(floatEq(*itPre, toAdd[2]));
-
-    // --- postfix operator++ ---
-    sparseVecFloat::const_iterator itPost = testSVF.begin();
-    sparseVecFloat::const_iterator itPostOld = itPost++;
-    assert(itPostOld.pos() == 0);
-    assert(itPost.pos() == 1);
-    assert(floatEq(*itPostOld, toAdd[0]));
-    assert(floatEq(*itPost, toAdd[1]));
-
-    // --- advance() ---
-    sparseVecFloat::const_iterator itAdv = testSVF.begin();
-    assert(floatEq(itAdv.value(), toAdd[0]));
-    bool stillValid = itAdv.advance();
-    assert(stillValid);
-    assert(itAdv.pos() == 1);
-    assert(floatEq(itAdv.value(), toAdd[1]));
-    stillValid = itAdv.advance();
-    assert(stillValid);
-    assert(itAdv.pos() == 2);
-    assert(floatEq(itAdv.value(), toAdd[2]));
-    bool pastEnd = itAdv.advance();
-    assert(!pastEnd);
-    assert(!itAdv.valid());
-
-    // --- go_to() ---
-    sparseVecFloat::const_iterator itGoto = testSVF.begin();
-    itGoto.go_to(2);
-    assert(itGoto.pos() == 2);
-    assert(floatEq(itGoto.value(), toAdd[2]));
-    itGoto.go_to(0);
-    assert(itGoto.pos() == 0);
-    assert(floatEq(itGoto.value(), toAdd[0]));
-    itGoto.go_to(1);
-    assert(itGoto.pos() == 1);
-    assert(floatEq(itGoto.value(), toAdd[1]));
-
-    // --- is_null() ---
-    sparseVecFloat::const_iterator itNull = testSVF.begin();
-    assert(!itNull.is_null());
-
-    // --- full iteration ---
-    int idx = 0;
-    for (auto it = testSVF.begin(); it != testSVF.end(); ++it, ++idx) {
-        assert(floatEq(*it, toAdd[idx]));
-    }
-    assert(idx == 3);
-}
-
-//----------------------------------------------------------------
-
-void SparseVecFloatImportTest(){
-    int N = 128000;
-    float m = 0.5f;
-    sparseVecFloat testSVF;
-    std::vector<float> temp(N*2);
-
-    for(int i = 0; i < N; i++){
-        temp[i] = (i * 0.001) * m;
-    }
-    for(int i = N; i < N*2; i++){
-        temp[i] = -1*(i * 0.001) * m;
-    }
-    
-    testSVF.import(temp.data(), N);
-    
-    int errorCount;
-    for(int i = 0; i < N; i++){
-        float err = std::fabs(temp[i] - testSVF.get(i));
-        if (err > 0.0f) {
-            errorCount++;
-        }
-    }
-    
-    assert(errorCount == 0);
-    BM_DECLARE_TEMP_BLOCK(tb)
-    testSVF.optimize(tb);
-    
-    errorCount = 0;
-    for(int i = 0; i < N; i++){
-        float err = std::fabs(temp[i] - testSVF.get(i));
-        if (err > 0.0f) {
-            errorCount++;
-        }
-    }
-    
-    assert(errorCount == 0);
-}
-
-//----------------------------------------------------------------
-
-void SparseVecFloatGeneralTests(){
-    auto floatEq = [](float a, float b) {
-        return std::fabs(a - b) < 0.001f;
-    };
-
-    float toAdd[] = {1.0123, -2.468, 340000.56};
-
-    sparseVecFloat testSVF;
-    
-    assert(testSVF.empty());
-
-    testSVF.push_back(toAdd[0]);
-
-    assert(!testSVF.empty());
-    assert(testSVF.size() == 1);
-    assert(floatEq(testSVF.get(0), toAdd[0]));
-
-    testSVF.push_back(toAdd[1]);
-    testSVF.push_back(toAdd[2]);
-
-    assert(testSVF.size() == 3);
-    assert(floatEq(testSVF.get(0), toAdd[0]));
-    assert(floatEq(testSVF.get(1), toAdd[1]));
-    assert(floatEq(testSVF.get(2), toAdd[2]));
-
-    sparseVecFloat testSVF2;
-    testSVF2.import(toAdd, 3);
-    assert(testSVF  == testSVF2);
-    assert(!(testSVF != testSVF2));
-
-    float toAdd2[] = {9.0f, -8.0f, -7.0f};
-    sparseVecFloat testSVF3;
-    testSVF3.import(toAdd2, 3);
-    assert(testSVF  != testSVF3);
-    assert(!(testSVF == testSVF3));
-
-    sparseVecFloat testSVFAssigned;
-    testSVFAssigned = testSVF;
-    assert(testSVFAssigned == testSVF);
-    assert(floatEq(testSVFAssigned.get(0), toAdd[0]));
-    assert(floatEq(testSVFAssigned.get(1), toAdd[1]));
-    assert(floatEq(testSVFAssigned.get(2), toAdd[2]));
-    assert(testSVFAssigned.size() == testSVF.size());
-
-    testSVF.set(1, 8.258f);
-    assert(!floatEq(testSVF.get(1), toAdd[1]));
-    assert(floatEq(testSVF.get(1), 8.258f));
-
-    testSVF.set(100, 100.001f);
-    assert(testSVF.size() == 101);
-    assert(floatEq(testSVF.get(100), 100.001f));
-    assert(floatEq(testSVF.get(50), 0.0f));
-
-    sparseVecFloat svA;
-    sparseVecFloat svB;
-    float aVals[] = {1.0f, -2.0f, 3.0f};
-    float bVals[] = {-4.0f, -5.0f, 6.0f};
-    svA.import(aVals, 3);
-    svB.import(bVals, 3);
-
-    svA.swap(svB);
-
-    assert(floatEq(svA.get(0), bVals[0]));
-    assert(floatEq(svA.get(1), bVals[1]));
-    assert(floatEq(svA.get(2), bVals[2]));
-    assert(floatEq(svB.get(0), aVals[0]));
-    assert(floatEq(svB.get(1), aVals[1]));
-    assert(floatEq(svB.get(2), aVals[2]));
-
-    svA.swap(svB);
-    assert(floatEq(svA.get(0), aVals[0]));
-    assert(floatEq(svA.get(1), aVals[1]));
-    assert(floatEq(svA.get(2), aVals[2]));
-}
-
-//----------------------------------------------------------------
-
-void SparseVecFloatSerializeTest(){
-    auto floatEq = [](float a, float b) {
-        return std::fabs(a - b) < 0.001f;
-    };
-
-    float toAdd[] = {1.0123, -2.468, 340000.56};
-
-    sparseVecFloat testSVF;
-    testSVF.import(toAdd, 3);
-    BM_DECLARE_TEMP_BLOCK(tb)
-    testSVF.optimize(tb);
-
-    bm::sparse_vector_float_serial_layout<sparseVecFloat> testLayout;
-    
-    bm::sparse_vector_float_serialize(testSVF, testLayout);
-
-    const unsigned char* buf = testLayout.buf();
-    bm::sparse_vector_float_deserialize(testSVF, buf);
-    
-    assert(testSVF.size() == 3);
-    assert(floatEq(testSVF.get(0), toAdd[0]));
-    assert(floatEq(testSVF.get(1), toAdd[1]));
-    assert(floatEq(testSVF.get(2), toAdd[2]));
-
-    sparseVecFloat testSVF2;
-    int N = 10000;
-    for(int i = 0; i < N; i++){
-        float f = i * 0.000123;
-        testSVF2.push_back(f);
-    }
-
-    testSVF2.optimize(tb);
-    bm::sparse_vector_float_serial_layout<sparseVecFloat> testLayout2;
-    bm::sparse_vector_float_serialize(testSVF2, testLayout2);
-
-    buf = testLayout2.buf();
-    bm::sparse_vector_float_deserializer<sparseVecFloat> testDeserializer;
-    sparseVecFloat testSVF2_restored;
-    testDeserializer.deserialize_range(testSVF2_restored, buf, 300, 400, true);
-    
-    int errorCount = 0;
-    for (int i = 300; i <= 400; i++) {
-        float f = i * 0.000123;
-        if (!floatEq(testSVF2_restored.get(i), f)){
-            errorCount++;
-        }
-    }
-    assert(errorCount == 0);
-
-    bm::bvector<> mask_bv;
-    int maskIndices[] = {0, 1, 50, 100, 500, 999, 5000, 9999};
-    int maskSize = sizeof(maskIndices) / sizeof(maskIndices[0]);
-    for (int i = 0; i < maskSize; i++)
-        mask_bv.set(maskIndices[i]);
-    
-    sparseVecFloat testSVF2_masked;
-    testDeserializer.deserialize(testSVF2_masked, buf, mask_bv);
-
-    errorCount = 0;
-    for (int i = 0; i < maskSize; i++) {
-        int idx   = maskIndices[i];
-        float f   = idx * 0.000123f;
-        if (!floatEq(testSVF2_masked.get(idx), f))
-            errorCount++;
-    }
-    assert(errorCount == 0);
-
-    assert(floatEq(testSVF2_masked.get(2),    0.0f));
-    assert(floatEq(testSVF2_masked.get(200),  0.0f));
-    assert(floatEq(testSVF2_masked.get(1000), 0.0f));
-}
-
-//----------------------------------------------------------------
-
-void SparseVecFloatRangeTests(){
-    auto floatEq = [](float a, float b) {
-        return std::fabs(a - b) < 0.001f;
-    };
-
-    float toAdd[] = {1.0123, -2.468, 340000.56, -7008.0, 0.900102};
-
-    sparseVecFloat testSVF;
-    testSVF.import(toAdd, 5);
-    BM_DECLARE_TEMP_BLOCK(tb)
-    testSVF.optimize(tb);
-
-    assert(testSVF.size() == 5);
-    testSVF.clear();
-    assert(testSVF.size() == 0);
-
-    testSVF.import(toAdd, 5);
-    testSVF.clear_range(1, 3);
-    assert(testSVF.size() == 5);
-    assert(floatEq(testSVF.get(0), toAdd[0]));
-    assert(floatEq(testSVF.get(1), 0.0));
-    assert(floatEq(testSVF.get(2), 0.0));
-    assert(floatEq(testSVF.get(3), 0.0));
-    assert(floatEq(testSVF.get(4), toAdd[4]));
-
-    testSVF.clear();
-    testSVF.import(toAdd, 5);
-
-    sparseVecFloat testSVF2(testSVF);
-    assert(testSVF.equal(testSVF2));
-
-    testSVF2.set(1, 0.0);
-    assert(!testSVF.equal(testSVF2));
-
-    assert(testSVF.compare(1, 2.468) == -1);
-    assert(testSVF.compare(1, -2.468) == 0);
-    assert(testSVF.compare(1, -3) == 1);
-
-    sparseVecFloat svf1;
-    sparseVecFloat svf2;
-    float toAdd1[] = {1.0123, -2.468, 0.0, 0.0, 0.0, 1.5};
-    float toAdd2[] = {0.0, 0.0, 0.0, -7008.0, 0.900102, 2.5};
-    svf1.import(toAdd1, 6);
-    svf2.import(toAdd2, 6);
-    svf1.optimize(tb);
-    svf2.optimize(tb);
-    
-    svf1.join(svf2);
-    assert(svf1.size() == 6);
-    assert(floatEq(svf1.get(0), toAdd[0]));
-    assert(floatEq(svf1.get(1), toAdd[1]));
-    assert(floatEq(svf1.get(2), 0.0));
-    assert(floatEq(svf1.get(3), toAdd[3]));
-    assert(floatEq(svf1.get(4), toAdd[4]));
-    assert(svf1.get(5) != svf1.get(5));
-
-    svf1.clear();
-    svf2.clear();
-    svf1.import(toAdd1, 6);
-    svf2.import(toAdd2, 6);
-    svf1.optimize(tb);
-    svf2.optimize(tb);
-
-    svf1.merge(svf2);
-    assert(svf1.size() == 6);
-    assert(floatEq(svf1.get(0), toAdd[0]));
-    assert(floatEq(svf1.get(1), toAdd[1]));
-    assert(floatEq(svf1.get(2), 0.0));
-    assert(floatEq(svf1.get(3), toAdd[3]));
-    assert(floatEq(svf1.get(4), toAdd[4]));
-    assert(svf1.get(5) != svf1.get(5));
-
-
-    svf1.clear();
-    svf2.clear();
-    svf1.import(toAdd1, 6);
-    svf2.import(toAdd2, 6);
-    svf1.optimize(tb);
-    svf2.optimize(tb);
-
-    svf1.copy_range(svf2, 2, 4);
-    assert(svf1.size() == 6);
-    assert(floatEq(svf1.get(0), 0.0));
-    assert(floatEq(svf1.get(1), 0.0));
-    assert(floatEq(svf1.get(2), 0.0));
-    assert(floatEq(svf1.get(3), toAdd[3]));
-    assert(floatEq(svf1.get(4), toAdd[4]));
-    assert(floatEq(svf1.get(5), 0.0));
-}
-
-//----------------------------------------------------------------
-
-void SparseVecFloatExtractionTests(){
-
-    auto floatEq = [](float a, float b) {
-        return std::fabs(a - b) < 0.001f;
-    };
-
-    int N = 128000;
-    float m = 0.5f;
-    sparseVecFloat testSVF;
-    std::vector<float> temp(N*2);
-
-    for(int i = 0; i < N; i++){
-        temp[i] = (i * 0.001) * m;
-    }
-    for(int i = N; i < N*2; i++){
-        temp[i] = -1*(i * 0.001) * m;
-    }
-
-    testSVF.import(temp.data(), N*2);
-    testSVF.optimize();
-
-    std::vector<float> testExtract(N*2);
-    testSVF.decode(testExtract.data(), 0, N*2);
-
-    int errorCount = 0;
-    for (int i = 0; i < N*2; i++) {
-        if (!floatEq(testExtract[i], temp[i])){
-            errorCount++;
-        }
-    }
-    assert(errorCount == 0);
-
-    
-    std::vector<float> testExtractRange(48000);
-    testSVF.extract_range(testExtractRange.data(), 48000, 16000);
-
-    errorCount = 0;
-    for (int i = 16000; i < 64000; i++) {
-        if (!floatEq(testExtractRange[i-16000], temp[i])){
-            errorCount++;
-        }
-    }
-    assert(errorCount == 0);
-
-    bm::id_t gatherIndeces[1024];
-    for(int i = 0; i < 1024; i++){
-        gatherIndeces[i] = rand() % 128000;
-    }
-
-    std::vector<float> testGather(1024);
-    testSVF.gather(testGather.data(), gatherIndeces, 1024, bm::BM_UNKNOWN);
-
-    errorCount = 0;
-    for (int i = 0; i < 1024; i++) {
-        if (!floatEq(testGather[i], temp[gatherIndeces[i]])){
-            errorCount++;
-        }
-    }
-    assert(errorCount == 0);
-}
-
-void SparseVecFloatBackInsertTests(){
-
-    auto floatEq = [](float a, float b) {
-        return std::fabs(a - b) < 0.001f;
-    };
-
-    sparseVecFloat testSVF;
-
-    sparseVecFloat::back_insert_iterator testBI(&testSVF);
-
-    testBI.add(1.0023);
-    assert(testSVF.size() == 1);
-    assert(floatEq(testSVF.get(0), 1.0023));
-
-    testBI.add(400005.6);
-    assert(testSVF.size() == 2);
-    assert(floatEq(testSVF.get(1), 400005.6));
-
-    sparseVecFloat::back_insert_iterator testBI2(testBI);
-    testBI2=78.9;
-    assert(testSVF.size() == 3);
-    assert(floatEq(testSVF.get(2), 78.9));
-
-    sparseVecFloat::back_insert_iterator testBI3(std::move(testBI));
-    testBI3=12345.6789;
-    assert(testSVF.size() == 4);
-    assert(floatEq(testSVF.get(3), 12345.6789));
-
-}
-
-//----------------------------------------------------------------
-//perf
-
-void SparseVecFloatStressTests(){
-    //Random data import and get test
-    {
-        std::random_device rd;
-        std::mt19937 gen(rd());
-        std::uniform_real_distribution<float> dist(-1000000.0f, 1000000.0f);
-
-        int N = 1000000;
-        std::vector<float> data(N);
-        for (int i = 0; i < N; i++)
-            data[i] = dist(gen);
-
-        sparseVecFloat sv;
-        sv.import(data.data(), N);
-
-        BM_DECLARE_TEMP_BLOCK(tb)
-        sv.optimize(tb);
-
-        int errorCount = 0;
-        for (int i = 0; i < N; i++) {
-            float restored = sv.get(i);
-            if (std::fabs(restored - data[i]) > 0.001f){
-                errorCount++;
-            }
-        }
-        assert(errorCount == 0);
-        assert((int)sv.size() == N);
-    }
-
-    //Random data push back and iterator tests
-    {
-        std::random_device rd;
-        std::mt19937 gen(rd());
-        std::uniform_real_distribution<float> dist(-500.0f, 500.0f);
-
-        int N = 500000;
-        std::vector<float> data(N);
-        sparseVecFloat sv;
-        for (int i = 0; i < N; i++){
-            data[i] = dist(gen);
-            sv.push_back(data[i]);
-        }
-
-        BM_DECLARE_TEMP_BLOCK(tb)
-        sv.optimize(tb);
-
-        // validate via iterator
-        int idx = 0;
-        int errorCount = 0;
-        for (auto it = sv.begin(); it != sv.end(); ++it, ++idx) {
-            if (std::fabs(*it - data[idx]) > 0.001f)
-                errorCount++;
-        }
-        assert(errorCount == 0);
-        assert(idx == N);
-    }
-
-    //Decaying Sinusoid Test
-    {
-        int N = 2000000;
-        std::vector<float> data(N);
-        const double e = std::exp(1.0);
-        for (int i = 0; i < N; i++) {
-            float t = 0.001f * i;
-            data[i] = 100.0f * (float)pow(e, -1.0 * t)
-                      * (cos(5.0f * t + 1.0f) + sin(5.0f * t + 1.0f));
-        }
-
-        sparseVecFloat sv;
-        sv.import(data.data(), N);
-        BM_DECLARE_TEMP_BLOCK(tb)
-        sv.optimize(tb);
-
-        int errorCount = 0;
-        for (int i = 0; i < N; i++) {
-            if (std::fabs(sv.get(i) - data[i]) > 0.001f)
-                errorCount++;
-        }
-        assert((int)sv.size() == N);
-    }
-
-}
-
-void SparseVecFloatSerialStressTests(){
-    //Serializing then deserializing a large random data set
-    {
-        std::random_device rd;
-        std::mt19937 gen(rd());
-        std::uniform_real_distribution<float> dist(-1000000.0f, 1000000.0f);
-
-        int N = 1000000;
-        std::vector<float> data(N);
-        for (int i = 0; i < N; i++)
-            data[i] = dist(gen);
-
-        sparseVecFloat sv;
-        sv.import(data.data(), N);
-        BM_DECLARE_TEMP_BLOCK(tb)
-        sv.optimize(tb);
-
-        bm::sparse_vector_float_serial_layout<sparseVecFloat> testLayout;
-        bm::sparse_vector_float_serialize(sv, testLayout);
-
-        // deserialize into a fresh vector
-        sparseVecFloat sv_restored;
-        const unsigned char* buf = testLayout.buf();
-        bm::sparse_vector_float_deserialize(sv_restored, buf);
-
-        // validate
-        assert((int)sv_restored.size() == N);
-        int errorCount = 0;
-        for (int i = 0; i < N; i++) {
-            if (std::fabs(sv_restored.get(i) - data[i]) > 0.001f)
-                errorCount++;
-        }
-        assert(errorCount == 0);
-    }
-
-    // Decaying Sinusoid Compression Test
-    {
-        int N = 2000000;
-        std::vector<float> data(N);
-        const double e = std::exp(1.0);
-        for (int i = 0; i < N; i++) {
-            float t = 0.001f * i;
-            data[i] = 100.0f * (float)pow(e, -1.0 * t)
-                      * (cos(5.0f * t + 1.0f) + sin(5.0f * t + 1.0f));
-        }
-
-        sparseVecFloat sv;
-        sv.import(data.data(), N);
-        BM_DECLARE_TEMP_BLOCK(tb)
-        sv.optimize(tb);
-
-        bm::sparse_vector_float_serial_layout<sparseVecFloat> testLayout;
-        bm::sparse_vector_float_serialize(sv, testLayout);
-
-        // deserialize into a fresh vector
-        sparseVecFloat sv_restored;
-        const unsigned char* buf = testLayout.buf();
-        bm::sparse_vector_float_deserialize(sv_restored, buf);
-
-        assert((int)sv_restored.size() == N);
-        int errorCount = 0;
-        for (int i = 0; i < N; i++) {
-            if (std::fabs(sv_restored.get(i) - data[i]) > 0.001f)
-                errorCount++;
-        }
-        assert(errorCount == 0);
-
-        // check that serialized size is smaller than raw size
-        size_t rawSize        = N * sizeof(float);
-        size_t serializedSize = testLayout.size();
-        assert(serializedSize < rawSize);
-    }
-}
-
-void SparseVecFloatTests(){
-
-    SparseVecFloatGeneralTests();
-    SparseVecFloatConstIteratorTests();
-    SparseVecFloatImportTest();
-    SparseVecFloatSerializeTest();
-    SparseVecFloatRangeTests();
-    SparseVecFloatExtractionTests();
-    SparseVecFloatBackInsertTests();
-    std::cout << "Sparse Vector Float Tests Complete" << std::endl;
-}
-
-struct SplitFloat {
-    unsigned int sign;
-    unsigned int exponent;
-    unsigned int mantissa;
-};
-
-SplitFloat split_float(float f) {
-    unsigned int bits;
-    std::memcpy(&bits, &f, sizeof(float));
-    return {
-        (bits >> 31) & 0x1,
-        (bits >> 23) & 0xFF,
-        bits         & 0x7FFFFF
-    };
-}
-
-void in_range(sparseVecFloat sv, float from, float to, sparseVecFloat::bvector_type &bv_out)
-{
-    bm::sparse_vector_scanner<sparseVecFloat> scan;
-    scan.find_range_float(sv, from, to, bv_out);
-}
-
-void in_range_vect(std::vector<float> fv, float from, float to, sparseVecFloat::bvector_type &bv_out)
-{
-    
-    for(int i = 0; i < fv.size(); i++){
-        if(fv[i] >= from && fv[i] <= to){
-            bv_out.set(i);
-        }
-    }
-}
-
-void in_range_const(sparseVecFloat sv, float from, float to, sparseVecFloat::bvector_type &bv_out)
-{
-    sparseVecFloat::const_iterator ci = sv.begin();
-    sparseVecFloat::const_iterator ciEnd = sv.end();
-
-    for (; ci != ciEnd; ci++){
-        float v = ci.value();
-        if(v >= from && v <= to){
-            bv_out.set(ci.pos());
-        }
-    }
-}
-
-void runTest(std::vector<float> temp, sparseVecFloat testSVF, float from, float to, int test){
-
-    bm::bvector<> bv_range;
-    bm::bvector<> bv_vector;
-    bm::bvector<> bv_const;
-
-    std::cout << "Test " << test << "[" << from << ", " << to << "]" << std::endl;
-
-    auto start = std::chrono::high_resolution_clock::now();
-    in_range(testSVF, from, to, bv_range);
-    auto end = std::chrono::high_resolution_clock::now();
-    auto duration = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
-    std::cout << "Scanner " << duration/1000 << "ms" << std::endl;
-
-    start = std::chrono::high_resolution_clock::now();
-    in_range_vect(temp, from, to, bv_vector);
-    end = std::chrono::high_resolution_clock::now();
-    duration = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
-    std::cout << "Vector " << duration/1000 << "ms" << std::endl;
-
-    start = std::chrono::high_resolution_clock::now();
-    in_range_const(testSVF, from, to, bv_const);
-    end = std::chrono::high_resolution_clock::now();
-    duration = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
-    std::cout << "Const Iterator " << duration/1000 << "ms" << std::endl;
-
-    bool range_eq_vector = (bv_range == bv_vector);
-    bool range_eq_const  = (bv_range == bv_const);
-
-    if (range_eq_vector && range_eq_const)
-    {
-        std::cout << "Test " << test << ": ALL 3 results match\n";
-    }
-    else
-    {
-        std::cout << "Test " << test << ": MISMATCH\n";
-        if (!range_eq_vector)
-        {
-            bm::bvector<> diff;
-            diff = bv_range ^ bv_vector;   // XOR shows differing bits
-            std::cout << "  range vs vector differs at " << diff.count() << " positions\n";
-            // print first few differing positions
-            auto en = diff.first();
-            for (int i = 0; i < 5 && en != diff.end(); ++i, ++en)
-                std::cout << "  position: " << *en << "\n";
-        }
-        if (!range_eq_const)
-        {
-            bm::bvector<> diff;
-            diff = bv_range ^ bv_const;
-            std::cout << "  range vs const differs at " << diff.count() << " positions\n";
-            auto en = diff.first();
-            for (int i = 0; i < 5 && en != diff.end(); ++i, ++en)
-                std::cout << "  position: " << *en << "\n";
-        }
-    }
-
-    std::cout << std::endl << std::endl;
-}
-
-void TestScannerLinear(){
-    BM_DECLARE_TEMP_BLOCK(tb)
-
-    int N = 20000000;
-    std::vector<float> temp(N);
-
-    for(int i = 0; i < N/2; i++)
-        temp[i] = -1.0f * i * 0.00123f;
-    for(int i = 0; i < N/2; i++)
-        temp[i+N/2] = i * 0.00123f;
-
-    sparseVecFloat testSVF;
-    testSVF.import(temp.data(), N);
-    testSVF.optimize(tb);    
-
-    // --- Test 1: large positive range ---
-    {
-        float from = 100.0f;
-        float to   = 500.0f;
-        runTest(temp, testSVF, from, to, 1);
-    }
-
-    // --- Test 2: small positive range ---
-    {
-        float from = 1.0f;
-        float to   = 2.0f;
-        runTest(temp, testSVF, from, to, 2);
-        
-    }
-
-    // --- Test 3: large negative range ---
-    {
-        float from = -500.0f;
-        float to   = -100.0f;
-        runTest(temp, testSVF, from, to, 3);
-    }
-
-    // --- Test 4: small negative range ---
-    {
-        float from = -2.0f;
-        float to   = -1.0f;
-        runTest(temp, testSVF, from, to, 4);
-    }
-
-    // --- Test 5: range crossing zero ---
-    {
-        float from = -10.0f;
-        float to   =  10.0f;
-        runTest(temp, testSVF, from, to, 5);
-    }
-
-    // --- Test 6: range with no results ---
-    {
-        float from = 99999.0f;
-        float to   = 99999.9f;
-        runTest(temp, testSVF, from, to, 6);
-    }
-
-    // --- Test 7: tiny range near zero ---
-    {
-        float from = -0.001f;
-        float to   =  0.001f;
-        runTest(temp, testSVF, from, to, 7);
-    }
-}
-
-void TestScannerRandom(){
-    BM_DECLARE_TEMP_BLOCK(tb)
-
-    int N = 20000000;
-    std::vector<float> temp(N);
-
-    std::random_device rd;
-    std::mt19937 gen(rd());
-
-    float upper = 1000000;
-    float lower = -1000000;
-    std::uniform_real_distribution<float> dis(lower, upper);
-
-    for (size_t i = 0; i < N; ++i) {
-        temp[i] = dis(gen);
-    }
-
-    sparseVecFloat testSVF;
-    testSVF.import(temp.data(), N);
-    testSVF.optimize(tb);
-
-    bm::bvector<> bv_out;
-
-    //Using a completely random dataset
-
-    // --- Test 1: large positive range ---
-    {
-        float from = 10000.0f;
-        float to   = 500000.0f;
-        runTest(temp, testSVF, from, to, 1);
-    }
-
-    // --- Test 2: small positive range ---
-    {
-        float from = 1.0f;
-        float to   = 2.0f;
-        runTest(temp, testSVF, from, to, 2);
-    }
-
-    // --- Test 3: large negative range ---
-    {
-        float from = -500000.0f;
-        float to   = -10000.0f;
-        runTest(temp, testSVF, from, to, 3);
-    }
-
-    // --- Test 4: small negative range ---
-    {
-        float from = -2.0f;
-        float to   = -1.0f;
-        runTest(temp, testSVF, from, to, 4);
-    }
-
-    // --- Test 5: range crossing zero ---
-    {
-        float from = -10000.0f;
-        float to   =  10000.0f;
-        runTest(temp, testSVF, from, to, 5);
-    }
-
-    // --- Test 6: range with no results ---
-    {
-        float from = 99999999.0f;
-        float to   = 99999999.9f;
-        runTest(temp, testSVF, from, to, 6);
-    }
-
-    // --- Test 7: tiny range near zero ---
-    {
-        float from = -0.001f;
-        float to   =  0.001f;
-        runTest(temp, testSVF, from, to, 7);
-    }
-}
-
-void TestScannerSkewed(){
-    BM_DECLARE_TEMP_BLOCK(tb)
-
-    int N = 20000000;
-    std::vector<float> temp(N);
-
-    std::random_device rd;
-    std::mt19937 gen(rd());
-
-    float upper = 1000000;
-    float lower = -1000000;
-    std::uniform_real_distribution<float> dis(lower, upper);
-
-    for (size_t i = 19000000; i < N; ++i) {
-        temp[i] = dis(gen);
-    }
-
-    sparseVecFloat testSVF;
-    testSVF.import(temp.data(), N);
-    testSVF.optimize(tb);
-
-    bm::bvector<> bv_out;
-
-    //Using a completely random dataset
-
-    // --- Test 1: large positive range ---
-    {
-        float from = 10000.0f;
-        float to   = 500000.0f;
-        runTest(temp, testSVF, from, to, 1);
-    }
-
-    // --- Test 2: small positive range ---
-    {
-        float from = 1.0f;
-        float to   = 2.0f;
-        runTest(temp, testSVF, from, to, 2);
-    }
-
-    // --- Test 3: large negative range ---
-    {
-        float from = -500000.0f;
-        float to   = -10000.0f;
-        runTest(temp, testSVF, from, to, 3);
-    }
-
-    // --- Test 4: small negative range ---
-    {
-        float from = -2.0f;
-        float to   = -1.0f;
-        runTest(temp, testSVF, from, to, 4);
-    }
-
-    // --- Test 5: range crossing zero ---
-    {
-        float from = -10000.0f;
-        float to   =  10000.0f;
-        runTest(temp, testSVF, from, to, 5);
-    }
-
-    // --- Test 6: range with no results ---
-    {
-        float from = 99999999.0f;
-        float to   = 99999999.9f;
-        runTest(temp, testSVF, from, to, 6);
-    }
-
-    // --- Test 7: tiny range near zero ---
-    {
-        float from = -0.001f;
-        float to   =  0.001f;
-        runTest(temp, testSVF, from, to, 7);
-    }
-}
-
-void TestScanner(){
-    std::cout << "Linear" << std::endl;
-    TestScannerLinear();
-    std::cout << "Random" << std::endl;
-    TestScannerRandom();
-    std::cout << "Skewed" << std::endl;
-    TestScannerSkewed();
-}
-
-int main(int argc, char *argv[]){
-    //SparseVecFloatTests();
-    //SparseVecFloatStressTests();
-    //SparseVecFloatSerialStressTests();
-    TestScanner();
-}
